// src/seri-repair/dto/create-seri-repair.dto.ts
export class CreateSeriRepairDto {
  locomotiveSeri: number;
  repairNumber: number;
  repairName: string;
  shortName: string;
  respondent: string;
  repairRate: number;
  stopMinute: number;
  priority: number;
  runBoolean: boolean;
  repairType: number;
  userInsertedId: number;
  userUpdatedId: number;
}